Added Result<_, _>

+++ firecracker-pilot/src/defaults.rs
@@ -41,7 +41,7 @@ pub const FIRECRACKER_FLAKE_DIR: &str =
     "/usr/share/flakes";
 pub const FIRECRACKER_VMID_DIR: &str =
     "/var/lib/firecracker/storage/tmp/flakes";
-pub const GC_THRESHOLD: i32 = 20;
+pub const GC_THRESHOLD: usize = 20;
 pub const VM_CID: u32 = 3;
 pub const VM_PORT: u32 =
     52;
